refactor: 로또 번호 String에서 로또 번호를 추출하는 DTO 수정
- 로또 번호를 구분자(", ")로 입력받아 로또 번호 Set을 반환해주는 DTO 사용
- 당첨번호는 이 클래스를 사용해서 얻어낸 로또번호 Set으로 Lotto 객체 생성해서 얻을 수 있다.
